done theme refactoring
