--- example/src/utils/themes/constants.tsx
@@ -1,3 +1,0 @@
-export const componentSizes = {
-  checkboxSize: 40,
-};
