+++ example/src/utils/themes/index.tsx
@@ -1,6 +1,2 @@
-export { spacing } from "./spacing";
 export { theme } from "./theme";
-export { typography } from "./typography";
 export { colors } from "./colors";
-export { componentSizes } from "./constants";
-export { typeface, regularType, boldType } from "./fonts";
