--- example/src/utils/themes/spacing.tsx
@@ -1,7 +1,0 @@
-export const spacing = {
-  xs: 8,
-  s: 16,
-  m: 24,
-  l: 32,
-  xl: 40,
-};
