+++ example/src/utils/themes/theme.tsx
@@ -1,22 +1,19 @@
-import { DarkTheme, DefaultTheme } from "@react-navigation/native";
+import { DefaultTheme, DarkTheme } from "@react-navigation/native";
 import { lightThemeColors, darkThemeColors } from "./colors";
-import { VariantTypes } from "./typography";
 
 export const theme = {
   light: {
     ...DefaultTheme,
     colors: {
       ...DefaultTheme.colors,
-      ...lightThemeColors.colors,
+      ...lightThemeColors,
     },
-    fonts: VariantTypes,
   },
   dark: {
     ...DarkTheme,
     colors: {
       ...DarkTheme.colors,
-      ...darkThemeColors.colors,
+      ...darkThemeColors,
     },
-    fonts: VariantTypes,
   },
 };
