+++ src/styles/themes/index.tsx
@@ -1,6 +1,5 @@
 export { spacing } from "./spacing";
 export { theme } from "./theme";
-export { typography } from "./typography";
 export { colors } from "./colors";
 export { componentSizes } from "./constants";
 export { typeface, regularType, boldType } from "./fonts";
